refactor in pkg
